Order Display working

+++ server.cpp
@@ -9,10 +9,10 @@ bool Server::fill_order_and_pay() {
 }
 bool Server::restock_and_pay() {
     _num_orders += 2;
-    return ((_num_orders % PAY_PERIOD) == 0) 
+    return ((_num_orders % PAY_PERIOD) == 0)
        || (((_num_orders - 1) % PAY_PERIOD) == 0);
 }
 double Server::pay_server() { }
 int Server::num_orders() {return _num_orders;}
-int Server::salary() {return _salary;}
+double Server::salary() {return _salary;}
 }
